+++ src/dom/element-container.ts
@@ -3,31 +3,32 @@ import {TextContainer} from './text-container';
 import {Bounds, parseBounds} from '../css/layout/bounds';
 import {isHTMLElementNode} from './node-parser';
 import {Context} from '../core/context';
+import {DebuggerType, isDebugging} from '../core/debugger';
 
 export const enum FLAGS {
     CREATES_STACKING_CONTEXT = 1 << 1,
     CREATES_REAL_STACKING_CONTEXT = 1 << 2,
-    IS_LIST_OWNER = 1 << 3
+    IS_LIST_OWNER = 1 << 3,
+    DEBUG_RENDER = 1 << 4
 }
 
 export class ElementContainer {
     readonly styles: CSSParsedDeclaration;
-    readonly textNodes: TextContainer[];
-    readonly elements: ElementContainer[];
+    readonly textNodes: TextContainer[] = [];
+    readonly elements: ElementContainer[] = [];
     bounds: Bounds;
-    flags: number;
+    flags = 0;
 
     constructor(protected readonly context: Context, element: Element) {
+        if (isDebugging(element, DebuggerType.PARSE)) {
+            debugger;
+        }
+
         this.styles = new CSSParsedDeclaration(context, window.getComputedStyle(element, null));
-        this.textNodes = [];
-        this.elements = [];
 
         if (isHTMLElementNode(element)) {
-            if (this.styles.animationDuration > 0) {
+            if (this.styles.animationDuration.some((duration) => duration > 0)) {
                 element.style.animationDuration = '0s';
-            }
-            if (this.styles.transitionDuration > 0) {
-                element.style.transitionDuration = '0s';
             }
 
             if (this.styles.transform !== null) {
@@ -37,6 +38,9 @@ export class ElementContainer {
         }
 
         this.bounds = parseBounds(this.context, element);
-        this.flags = 0;
+
+        if (isDebugging(element, DebuggerType.RENDER)) {
+            this.flags |= FLAGS.DEBUG_RENDER;
+        }
     }
 }